requirements.txt
fastapi
uvicorn
python-multipart
yara-python
python-evtx
pytest
sqlalchemy
bcrypt
